// src/modules/company/enums/messages.enum.ts
export enum CompanyMessage {
  CREATED = 'Company successfully created.',
  RETRIEVED_ALL = 'List of all companies retrieved successfully.',
  RETRIEVED_ONE = 'Company retrieved successfully.',
  UPDATED = 'Company updated successfully.',
  DELETED = 'Company deleted successfully.',
  NOT_FOUND = 'Company with the specified ID was not found.',
}